requirements.txt
# common Dependencies
pyYaml

# For python Script
requests

# For Ansible Playbook
ansible